[clang][cas] Create explicit IncludeTree::ModuleMap
Instead of putting the modulemap file in the include-tree filesystem and
parsing it at build time, create a data structure that represents just
the parts of the modulemap we need when building the module:

* Flags (explicit, system, framework, etc.)
* Exports (export *, export Foo)
* LinkLibraries (implicit framework autolink, link Foo)

Additionally, we add modular headers lazily by inserting known headers
when we encounter an include-tree header that is part of a submodule
(this is needed for missing #include diagnostics).

This removes the possibility of mismatches between header paths seen
during modulemap parsing from the paths we embed due to #includes, for
exmaple due to VFS virtual vs external paths.